+++ tests/Unit/PersonSearchResultNormalizerTest.php
@@ -117,4 +117,86 @@ final class PersonSearchResultNormalizerTest extends TestCase
         self::assertSame('035-1234567', $normalized[0]['phone']);
         self::assertSame('maria.jansen@example.org', $normalized[0]['email']);
     }
+
+    public function testNormalizeDetailPersonMapsPpaPersonPayloadToKiwiCustomerModel(): void
+    {
+        $credential = new HupApiCredential(
+            name: 'tvk',
+            title: 'TV Krant',
+            mandant: 'HMC',
+            supportsPersonLookup: true,
+            username: 'tvk-user',
+            password: 'tvk-password',
+            refreshToken: null,
+        );
+        $rawPerson = [
+            'rId' => '11860448',
+            'personNumber' => '41929371',
+            'division' => [
+                'rId' => '14',
+            ],
+            'lastName' => 'Bakker-Bakker',
+            'surName' => 'de',
+            'birthDay' => '1945-11-20',
+            'matchCode' => '7831CM33',
+            'initials' => 'R.',
+            'addressType' => [
+                'name' => 'HEER',
+            ],
+            'contacts' => [
+                'emails' => [
+                    ['emailAddress' => 'wiesje_meeringa@hotmail.nl'],
+                ],
+                'phones' => [
+                    ['number' => '0591522006'],
+                ],
+                'addresses' => [
+                    [
+                        'address' => [
+                            'street' => 'Ph. Lindemanstraat',
+                            'postCode' => '7831CM',
+                            'city' => 'NIEUW-WEERDINGE',
+                            'housenumber' => [
+                                'housenumber' => '33',
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+            'payments' => [
+                'ibanItems' => [
+                    ['iban' => 'NL80INGB0001340187'],
+                ],
+            ],
+            'references' => [
+                [
+                    'origin' => 'UIS',
+                    'identifier' => '41929371',
+                ],
+            ],
+        ];
+
+        $normalizer = new PersonSearchResultNormalizer();
+
+        $normalized = $normalizer->normalizeDetailPerson($rawPerson, $credential, '11860448');
+
+        self::assertSame(11860448, $normalized['id']);
+        self::assertSame('11860448', $normalized['personId']);
+        self::assertSame('41929371', $normalized['personNumber']);
+        self::assertSame('', $normalized['firstName']);
+        self::assertSame('de', $normalized['middleName']);
+        self::assertSame('Bakker-Bakker', $normalized['lastName']);
+        self::assertSame('Dhr.', $normalized['salutation']);
+        self::assertSame('1945-11-20', $normalized['birthday']);
+        self::assertSame('Ph. Lindemanstraat 33', $normalized['address']);
+        self::assertSame('7831CM', $normalized['postalCode']);
+        self::assertSame('33', $normalized['houseNumber']);
+        self::assertSame('NIEUW-WEERDINGE', $normalized['city']);
+        self::assertSame('wiesje_meeringa@hotmail.nl', $normalized['email']);
+        self::assertSame('0591522006', $normalized['phone']);
+        self::assertSame('HMC', $normalized['mandant']);
+        self::assertSame('14', $normalized['divisionId']);
+        self::assertSame('NL80INGB0001340187', $normalized['iban']);
+        self::assertSame([['origin' => 'UIS', 'identifier' => '41929371']], $normalized['references']);
+    }
 }